requirements.txt
# SQLAlchemy — используется для управления моделями и CRUD через Python, с возможностью raw SQL там, где нужна оптимизация
sqlalchemy==2.0.49

# python-dotenv — загружает конфиги из .env для настройки подключения к БД и Redis
python-dotenv==1.2.2

# Pydantic — схемы запросов и ответов FastAPI, валидация данных на уровне API
pydantic==2.12.5

# Redis — кэширование популярных постов, ускоряет GET-запросы и снижает нагрузку на PostgreSQL
redis==7.4.0

# FastAPI — основной API фреймворк проекта, асинхронные эндпоинты и автогенерация документации
fastapi==0.135.3

# Uvicorn — сервер для FastAPI, обеспечивает асинхронное выполнение и поддержку live-reload в разработке
uvicorn==0.44.0

# pytest — тестирование бизнес-логики и кэширования, запуск интеграционных проверок
pytest==9.0.3

# httpx — имитация HTTP-запросов в тестах, проверка API без реального клиента
httpx==0.28.1

# psycopg2-binary — подключение к PostgreSQL для SQLAlchemy и raw SQL
psycopg2-binary==2.9.11